Updated migrations to not error out if tables exist

+++ src/database/migrations/2016_10_11_141419_create_blog_table.php
@@ -13,21 +13,23 @@ class CreateBlogTable extends Migration
      */
     public function up()
     {
-        Schema::create('blogs', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
-            $table->string('title');
-            $table->string('slug');
-            $table->boolean('published')->default(1);
-            $table->text('body');
-            $table->timestamp('published_at');
-            $table->timestamps();
-            
-            $table->foreign('user_id')
-            ->references('id')
-            ->on('users')
-            ->onDelete('cascade');
-        });
+        if (!Schema::hasTable('blogs')) {
+            Schema::create('blogs', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('user_id')->unsigned();
+                $table->string('title');
+                $table->string('slug');
+                $table->boolean('published')->default(1);
+                $table->text('body');
+                $table->timestamp('published_at');
+                $table->timestamps();
+
+                $table->foreign('user_id')
+                    ->references('id')
+                    ->on('users')
+                    ->onDelete('cascade');
+            })
+        }
     }
 
     /**
